==> SCRIPTS/Proteomics Analysis/03_Volcano_plots_Proteomics_Fig1 Biomarkers_All_timepts_SJ.Rmd <==
---
title: "PF_Proteomics_Figure 1 Biomarkers correlation"
author: 'Volcano plots - All timepoints'
date: "SJ"
output:
  html_document:
    toc: yes
    number_sections: true
    toc_depth: 5
  pdf_document:
    toc: yes
    number_sections: true
    toc_depth: 5
urlcolor: blue
linkcolor: red
---

```{r global_options, include=FALSE}
rm(list = ls())
# if (!requireNamespace("BiocManager", quietly = TRUE))
#     install.packages("BiocManager")
# BiocManager::install("EnhancedVolcano")
library(EnhancedVolcano)
library(dplyr)
library(forecast)
library(kableExtra)
library(readxl) # Input file is xlsx (not csv)
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, comment=NA, warning=FALSE,
                      fig.align = 'center',fig.width = 6, fig.asp = 0.8, 
                      out.width = "80%")
#knitr::opts_knit$set(root.dir = "G:/My Drive/Prolonged_Fasting_Study")
knitr::opts_knit$set(root.dir = "../..")
theme_update(plot.title = element_text(hjust = 0.5)) # center plots
```


# Useful functions

```{r, message=FALSE}
source("./Functions/Metabolomics functions_SJ_GF.R")
source("./Functions/Correlation_functions.R")
```

# Working directory and datasets
```{r}
load("./SCRIPTS/RDA_Non_imputed_dfs_SJ.rda")
setwd("./DATA/PROLONGED FASTING/GTAC_somascan/PF_Somalogic_18-045")
df_proteomics=read_excel("Cleaned_Proteomics data_AM.xlsx",sheet = "Sheet2_SJedit")
```

# Format dataset structure

```{r}
# Clean proteomics data
### !! Only run following code altogether
df_proteomics = data.frame(Timept_ID=paste(df_proteomics$ID,
                                           df_proteomics$Timepoint,sep = "."),
                           df_proteomics[,-c(1:5)])

# Transposed for convenient merging on Timepoint+ID (not used here)
Timept_ID=df_proteomics$Timept_ID
df_proteomics<- data.frame(t(df_proteomics[,-1]))
colnames(df_proteomics) <- Timept_ID

# Get named group factor
group<-factor(sub(".*\\.", "", colnames(df_proteomics)))
names(group)<-colnames(df_proteomics)

# Normalise each row (in case of skews)
df_proteomics<-data.frame(t(apply(df_proteomics,1,BoxCoxTransf)))

# Biomarkers
### !! Only run following code altogether
df_main = data.frame(Timept_ID=paste(df_fast_wide$ID,df_fast_wide$Timepoint,sep = "."),
                     df_fast_wide[,-c(1,2)])

# Transposed for convenient merging on Timepoint+ID (not used here)
Timept_ID=df_main$Timept_ID
df_main<- data.frame(t(df_main[,-1]))
colnames(df_main) <- Timept_ID

(NAs_cols=which(colMeans(is.na(df_main))!=0)) # Bililrubin, UrineSG, U_ketones, U_pH removed
colSums(is.na(df_main[,NAs_cols]))
df_main=df_main[-c(22,33,55,57),]
###

# Normalise each row (in case of skews)
df_main<-data.frame(t(apply(df_main,1,BoxCoxTransf)))
```

# Correlation of Figure 1 Biomarkers with Metabolites

## Panel A

### BHB

```{r}
df_merged<-func_merge_metabolomics_biomarker(df_proteomics,df_main, Beta.hydroxy.butyrate)
rownames(df_merged)[rownames(df_merged)=="biomarker"] <- "BHB"
```

#### Correlation results table

```{r}
corr_table = func_corr_results_table(df_merged)
corr_bhb_all<-corr_table[abs(corr_table$FDR)<=0.2,] 
corr_bhb_all %>% func_Kable()

corr_bhb_all_pearsons<-data.frame(cor_pearsons=corr_bhb_all$cor_pearsons)
rownames(corr_bhb_all_pearsons)<-rownames(corr_bhb_all)
corr_bhb_all<-rownames(corr_bhb_all)
```

#### Volcano Plot

```{r}
func_volcano_plot(corr_table, biomarker = "BHB", main_title="BHB Volcano Plot")
```

### Cortisol

```{r}
df_merged<-func_merge_metabolomics_biomarker(df_proteomics,df_main, Cortisol)
rownames(df_merged)[rownames(df_merged)=="biomarker"] <- "Cortisol"
```

#### Correlation results table

```{r}
corr_table = func_corr_results_table(df_merged)
corr_cort_all <-corr_table[abs(corr_table$FDR)<=0.2,] 
corr_cort_all %>% func_Kable()

corr_cort_all_pearsons<-data.frame(cor_pearsons=corr_cort_all$cor_pearsons)
rownames(corr_cort_all_pearsons)<-rownames(corr_cort_all)
corr_cort_all<-rownames(corr_cort_all)
```

#### Volcano Plot

```{r}
func_volcano_plot(corr_table, biomarker = "Cortisol", main_title="Cortisol Volcano Plot")
```

### IGF.1

```{r}
df_merged<-func_merge_metabolomics_biomarker(df_proteomics,df_main, IGF.1)
rownames(df_merged)[rownames(df_merged)=="biomarker"] <- "IGF.1"
```

#### Correlation results table

```{r}
corr_table = func_corr_results_table(df_merged)
corr_igf1_all<-corr_table[abs(corr_table$FDR)<=0.2,] 
corr_igf1_all %>% func_Kable()

corr_igf1_all_pearsons<-data.frame(cor_pearsons=corr_igf1_all$cor_pearsons)
rownames(corr_igf1_all_pearsons)<-rownames(corr_igf1_all)
corr_igf1_all<-rownames(corr_igf1_all)
```

#### Volcano Plot

```{r}
func_volcano_plot(corr_table, biomarker = "IGF.1", main_title="IGF.1 Volcano Plot")
```

### FreeT3

```{r}
df_merged<-func_merge_metabolomics_biomarker(df_proteomics,df_main, FreeT3)
rownames(df_merged)[rownames(df_merged)=="biomarker"] <- "FreeT3"
```

#### Correlation results table

```{r}
corr_table = func_corr_results_table(df_merged)
corr_t3_all<-corr_table[abs(corr_table$FDR)<=0.2,]
corr_t3_all %>% func_Kable()

corr_t3_all_pearsons<-data.frame(cor_pearsons=corr_t3_all$cor_pearsons)
rownames(corr_t3_all_pearsons)<-rownames(corr_t3_all)
corr_t3_all<-rownames(corr_t3_all)
```

#### Volcano Plot

```{r}
func_volcano_plot(corr_table, biomarker = "FreeT3", main_title="FreeT3 Volcano Plot")
```

## Panel B

### Glucose

```{r}
df_merged<-func_merge_metabolomics_biomarker(df_proteomics,df_main, Glucose)
rownames(df_merged)[rownames(df_merged)=="biomarker"] <- "Glucose"
```

#### Correlation results table

```{r}
corr_table = func_corr_results_table(df_merged)
corr_gluc_all<-corr_table[abs(corr_table$FDR)<=0.2,]
corr_gluc_all %>% func_Kable()

corr_gluc_all_pearsons<-data.frame(cor_pearsons=corr_gluc_all$cor_pearsons)
rownames(corr_gluc_all_pearsons)<-rownames(corr_gluc_all)
corr_gluc_all<-rownames(corr_gluc_all)
```

#### Volcano Plot

```{r}
func_volcano_plot(corr_table, biomarker = "Glucose", main_title="Glucose Volcano Plot")
```

### Insulin

```{r}
df_merged<-func_merge_metabolomics_biomarker(df_proteomics,df_main, Insulin)
rownames(df_merged)[rownames(df_merged)=="biomarker"] <- "Insulin_biomarker"
```

#### Correlation results table

```{r}
corr_table = func_corr_results_table(df_merged)
corr_ins_all<-corr_table[abs(corr_table$FDR)<=0.2,] 
corr_ins_all %>% func_Kable()

corr_ins_all_pearsons<-data.frame(cor_pearsons=corr_ins_all$cor_pearsons)
rownames(corr_ins_all_pearsons)<-rownames(corr_ins_all)
corr_ins_all<-rownames(corr_ins_all)
```

#### Volcano Plot

```{r}
func_volcano_plot(corr_table, biomarker = "Insulin_biomarker", main_title="Insulin Volcano Plot")
```

### Non_HDL_Chol

```{r}
df_merged<-func_merge_metabolomics_biomarker(df_proteomics,df_main, Non_HDL_Chol)
rownames(df_merged)[rownames(df_merged)=="biomarker"] <- "Non_HDL_Chol"
```

#### Correlation results table

```{r}
corr_table = func_corr_results_table(df_merged)
corr_nonhdlchol_all<-corr_table[abs(corr_table$FDR)<=0.2,]
corr_nonhdlchol_all %>% func_Kable()

corr_nonhdlchol_all_pearsons<-data.frame(cor_pearsons=corr_nonhdlchol_all$cor_pearsons)
rownames(corr_nonhdlchol_all_pearsons)<-rownames(corr_nonhdlchol_all)
corr_nonhdlchol_all<-rownames(corr_nonhdlchol_all)
```

#### Volcano Plot

```{r}
func_volcano_plot(corr_table, biomarker = "Non_HDL_Chol", main_title="Non_HDL_Chol Volcano Plot")
```

### Triglycerides

```{r}
df_merged<-func_merge_metabolomics_biomarker(df_proteomics,df_main, Triglycerides)
rownames(df_merged)[rownames(df_merged)=="biomarker"] <- "Triglycerides"
```

#### Correlation results table

```{r}
corr_table = func_corr_results_table(df_merged)
corr_tg_all<-corr_table[abs(corr_table$FDR)<=0.2,] 
corr_tg_all %>% func_Kable()

corr_tg_all_pearsons<-data.frame(cor_pearsons=corr_tg_all$cor_pearsons)
rownames(corr_tg_all_pearsons)<-rownames(corr_tg_all)
corr_tg_all<-rownames(corr_tg_all)
```

#### Volcano Plot

```{r}
func_volcano_plot(corr_table, biomarker = "Triglycerides", main_title="Triglycerides Volcano Plot")
```

# Intersections/Common Proteins (highly correlated) in Biomarkers 

```{r}
# All
Reduce(intersect,list(corr_bhb_all, corr_cort_all, corr_igf1_all, corr_t3_all,
                      corr_gluc_all, corr_ins_all, corr_nonhdlchol_all, corr_tg_all))

# Panel A
Reduce(intersect,list(corr_bhb_all, corr_cort_all, corr_igf1_all, corr_t3_all))
Reduce(intersect,list(corr_bhb_all, corr_cort_all, corr_t3_all))
Reduce(intersect,list(corr_cort_all, corr_t3_all))
Reduce(intersect,list(corr_bhb_all, corr_t3_all))
Reduce(intersect,list(corr_bhb_all, corr_cort_all))

# Panel B
Reduce(intersect,list(corr_gluc_all, corr_ins_all, corr_nonhdlchol_all, corr_tg_all))

Reduce(intersect,list(corr_gluc_all, corr_ins_all))

# Note: Except for tg, no other biomarker shows commonality with nonhdlchol.
# Vice-versa is also true - Except for nonhdl chol, no other has common with tg.
Reduce(intersect,list(corr_nonhdlchol_all, corr_tg_all))

# Mix
Reduce(intersect,list(corr_t3_all, corr_ins_all))
Reduce(intersect,list(corr_igf1_all, corr_ins_all)) # nothing in common
Reduce(intersect,list(corr_bhb_all, corr_gluc_all))

# Cort, igf1 ,freet3, bhb, ins
```

# Save Significant Correlated Metabolites in RDA

```{r}
# Save sig proteins correlated (all timepts)
setwd("./SCRIPTS/Proteomics Analysis")
# cor_protein_lists<-mget(ls(pattern = "_all$")) # list2env(cor_protein_lists, .GlobalEnv) # to expand elements compressed in list.
save(corr_bhb_all,corr_cort_all,corr_igf1_all, corr_t3_all,
     corr_gluc_all,corr_ins_all,corr_nonhdlchol_all,corr_tg_all,
     file="RDA_All_timepts_Correlation_Proteins_Biomarkers.rda")

cor_tables <- mget(ls(pattern = "all_pearsons$"))
save(cor_tables, file = "RDA_All_timepts_Corr_tables_Proteins_Biomarkers.rda")
```
