==> SCRIPTS/Proteomics Analysis/04_Proteomics_Pathway_Enrichment Analysis_SJ.Rmd <==
---
title: "PF_Proteomics Pathway Enrichment Analysis"
author: 'Uniprot, Annotated Proteins'
date: "SJ"
output:
  html_document:
    toc: yes
    number_sections: true
    toc_depth: 5
  pdf_document:
    toc: yes
    number_sections: true
    toc_depth: 5
urlcolor: blue
linkcolor: red
---

```{r global_options, include=FALSE}
rm(list = ls())
# if (!requireNamespace("BiocManager", quietly = TRUE))
#     install.packages("BiocManager")
# BiocManager::install("EnhancedVolcano")
library(EnhancedVolcano)
library(dplyr)
library(forecast)
library(VennDiagram)
library(readxl) # Input file is xlsx (not csv)
library(UniprotR)
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, comment=NA, warning=FALSE,
                      fig.width = 6, fig.asp = 0.8, out.width = "80%")
#knitr::opts_knit$set(root.dir = "G:/My Drive/Prolonged_Fasting_Study")
knitr::opts_knit$set(root.dir = "../..")
theme_update(plot.title = element_text(hjust = 0.5)) # center plots
```


# Useful functions for getting list results
```{r, message=FALSE}
source("./Functions/Metabolomics functions_SJ_GF.R")
source("./Functions/Correlation_functions.R")

# Inputs protein short forms (with or w/o dots) to return Accs_no.
# If uniprot_target_name = T --> using Target column
# Defaults to Target_dots column
func_Accs_no_proteins <- function(protein_list, df_UniProt_IDs, uniprot_target_name=F){
  prot_list<-func_prots_X_numeric(protein_list)
  if (uniprot_target_name){
    accs=na.omit(df_UniProt_IDs[match(prot_list,df_UniProt_IDs$Target),]$UniProt)
  } else {
    accs<-na.omit(df_UniProt_IDs[match(prot_list,df_UniProt_IDs$Target_dots),]$UniProt)
  }
  return(accs)
}

# To select pathway analysis required (options: Pathway.Enr, BP, CC, MF)
func_path_enr <- function(protein_Accs_list, type = "Path.Enr"){
  if (type == "Path.Enr"){
    res <- Pathway.Enr(protein_Accs_list)
  } else if (type == "BP"){
    res<-Enrichment.BP(protein_Accs_list)
  } else if (type == "CC"){
    res<-Enrichment.CC(protein_Accs_list)
  } else if (type == "MF"){
    res<-Enrichment.MF(protein_Accs_list)
  }
  return (res)
}

# Returns pvalue, pathway, common proteins belonging to pathway for Enrichment.obj
func_extract_relevant_paths<- function(enr.obj){
  data_enr.obj<-enr.obj$data
  res<-data.frame(p_val = data_enr.obj$p_value,
                  pathway = data_enr.obj$term_name,
                  common_proteins = data_enr.obj$intersection)
  return (res)
}

# Removes X as prefix if 2nd character is numeric.
func_prots_X_numeric <- function(protein_list){
  for (i in 1:length(protein_list)){
  list_chars<-unlist(strsplit(protein_list[i],""))
  # Take proteins that start with X and followed by a number, removes X in front.
  if (list_chars[1]=="X" & is.numeric(as.numeric(list_chars[2])) & 
      !is.na(as.numeric(list_chars[2]))){
    protein_list[i] <- paste(list_chars[-1],collapse = "")
    }
  }
  return (protein_list)
}
```

# Datasets
```{r}
setwd("./DATA/PROLONGED FASTING/GTAC_somascan/PF_Somalogic_18-045")
df_proteomics=data.frame(read_excel("Cleaned_Proteomics data_AM.xlsx",sheet = "Sheet2_SJedit"))

# Sheet that has Protein name with Xs and dots..
# NOTE: before importing, convert all ",", "+", ":", " ", "-", other punctuation marks to "." in Target_dots column (C). This is important so as to not lead to NAs when merging.
# NOTE: X before numeric value in name taken care of through func_prots_X_numeric
df_uniprotIDs=data.frame(read_excel("Cleaned_Proteomics data_AM.xlsx",sheet = "Uniprot_dict"))
```

# Loaded FC and Correlation Tables

```{r}
setwd("./SCRIPTS/Proteomics Analysis")

load("RDA_Proteins_FC.rda") # Fold changes in proteins
# load("RDA_BLEFER_Correlation_Proteins_Biomarkers.rda") # Correlation with timepoints - only BL, EF, ER. 

load("RDA_All_timepts_Correlation_Proteins_Biomarkers.rda") #Same correlations but with all timepts.

# Note: If W1 included in BLEFER RDA then results almost identical to All-timepts rda.
# Note: If you want to work with W1, then load all-timepts instead of BLEFER.
# Note: BLEFER_corr = corr_biomarker; All-timepts_corr = corr_biomarker_all
```

# Pathway Enrichment

## Fold-changes

### BLEF_BLW1 Common Proteins

```{r}
a<-func_Accs_no_proteins(blef_blw1_intrsct,df_uniprotIDs)

enr_blef_blw1_bp<-func_path_enr(a,"BP") 
enr_blef_blw1_cc<-func_path_enr(a,"CC") 
enr_blef_blw1_mf<-func_path_enr(a,"MF")

paths_blef_blw1_bp <- func_extract_relevant_paths(enr_blef_blw1_bp)
paths_blef_blw1_cc <- func_extract_relevant_paths(enr_blef_blw1_cc)
paths_blef_blw1_mf <- func_extract_relevant_paths(enr_blef_blw1_mf)

# Gene Ontology
taxa_names<-GetNamesTaxa(a)
PlotGenesNetwork(taxa_names) # radial interactive proteins from genes (if grouping exists?)
PlotEnrichedGO(a) # BP, MF, CC (sorted and selected by p-value)

protein_info<-GetProteinGOInfo(a)
PlotGOAll(protein_info) # BP, MF, CC (sorted by count)
PlotGoInfo(protein_info) # Table formate of GOAll
```

### BLEF_EFER Common Proteins

```{r}
a<-func_Accs_no_proteins(blef_efer_intrsct,df_uniprotIDs)


enr_blef_efer_bp<-func_path_enr(a,"BP")
enr_blef_efer_cc<-func_path_enr(a,"CC")
enr_blef_efer_mf<-func_path_enr(a,"MF")

paths_blef_efer_bp <- func_extract_relevant_paths(enr_blef_efer_bp)
paths_blef_efer_cc <- func_extract_relevant_paths(enr_blef_efer_cc)
paths_blef_efer_mf <- func_extract_relevant_paths(enr_blef_efer_mf)

# Gene Ontology
taxa_names<-GetNamesTaxa(a)
PlotGenesNetwork(taxa_names)
PlotEnrichedGO(a) # BP, MF, CC (p-value sorted)

protein_info<-GetProteinGOInfo(a)
PlotGOAll(protein_info) # BP, MF, CC (count-sorted)
PlotGoInfo(protein_info)

PlotEnrichedPathways(a)
```

## Correlations

### Panel A

#### BHB

```{r}
a<-func_Accs_no_proteins(corr_bhb_all,df_uniprotIDs)

enr_bhb_corr_path<-func_path_enr(a) # Pathway.Enr -> REAC and KEGG
enr_bhb_corr_bp<-func_path_enr(a,"BP")
enr_bhb_corr_cc<-func_path_enr(a,"CC")
enr_bhb_corr_mf<-func_path_enr(a,"MF")

paths_bhb_corr_path.enr <- func_extract_relevant_paths(enr_bhb_corr_path)
paths_bhb_corr_bp <- func_extract_relevant_paths(enr_bhb_corr_bp)
paths_bhb_corr_cc <- func_extract_relevant_paths(enr_bhb_corr_cc)
paths_bhb_corr_mf <- func_extract_relevant_paths(enr_bhb_corr_mf)

# Gene Ontology
taxa_names<-GetNamesTaxa(a)
PlotGenesNetwork(taxa_names)
PlotEnrichedGO(a)

protein_info<-GetProteinGOInfo(a)
PlotGOAll(protein_info)
PlotGoInfo(protein_info)

PlotEnrichedPathways(a)

```

```{r}
print("Early termination of code..")
knitr::knit_exit()
```

#### Cortisol

```{r}
a<-func_Accs_no_proteins(corr_cort_all,df_uniprotIDs)

enr_cort_corr_path<-func_path_enr(a)
enr_cort_corr_bp<-func_path_enr(a,"BP")
enr_cort_corr_cc<-func_path_enr(a,"CC")
enr_cort_corr_mf<-func_path_enr(a,"MF")

paths_cort_corr_path.enr <- func_extract_relevant_paths(enr_cort_corr_path)
paths_cort_corr_bp <- func_extract_relevant_paths(enr_cort_corr_bp)
paths_cort_corr_cc <- func_extract_relevant_paths(enr_cort_corr_cc)
paths_cort_corr_mf <- func_extract_relevant_paths(enr_cort_corr_mf)

# Gene Ontology
taxa_names<-GetNamesTaxa(a)
PlotGenesNetwork(taxa_names)
PlotEnrichedGO(a)

protein_info<-GetProteinGOInfo(a)
PlotGOAll(protein_info)
PlotGoInfo(protein_info)

PlotEnrichedPathways(a)
```

#### IGF1

```{r}
a<-func_Accs_no_proteins(corr_igf1_all,df_uniprotIDs)

enr_igf1_corr_path<-func_path_enr(a)
enr_igf1_corr_bp<-func_path_enr(a,"BP")
enr_igf1_corr_cc<-func_path_enr(a,"CC")
enr_igf1_corr_mf<-func_path_enr(a,"MF")

paths_igf1_corr_path.enr <- func_extract_relevant_paths(enr_igf1_corr_path)
paths_igf1_corr_bp <- func_extract_relevant_paths(enr_igf1_corr_bp)
paths_igf1_corr_cc <- func_extract_relevant_paths(enr_igf1_corr_cc)
paths_igf1_corr_mf <- func_extract_relevant_paths(enr_igf1_corr_mf)

# Gene Ontology
taxa_names<-GetNamesTaxa(a)
PlotGenesNetwork(taxa_names)
PlotEnrichedGO(a)

protein_info<-GetProteinGOInfo(a)
PlotGOAll(protein_info)
PlotGoInfo(protein_info)

PlotEnrichedPathways(a)
```

#### FreeT3

```{r}
a<-func_Accs_no_proteins(corr_t3_all,df_uniprotIDs)

enr_t3_corr_path<-func_path_enr(a)
enr_t3_corr_bp<-func_path_enr(a,"BP")
enr_t3_corr_cc<-func_path_enr(a,"CC")
enr_t3_corr_mf<-func_path_enr(a,"MF")

paths_t3_corr_path.enr <- func_extract_relevant_paths(enr_t3_corr_path)
paths_t3_corr_bp <- func_extract_relevant_paths(enr_t3_corr_bp)
paths_t3_corr_cc <- func_extract_relevant_paths(enr_t3_corr_cc)
paths_t3_corr_mf <- func_extract_relevant_paths(enr_t3_corr_mf)

# Gene Ontology
taxa_names<-GetNamesTaxa(a)
PlotGenesNetwork(taxa_names)
PlotEnrichedGO(a)

protein_info<-GetProteinGOInfo(a)
PlotGOAll(protein_info)
PlotGoInfo(protein_info)

PlotEnrichedPathways(a)
```

### Panel B

#### Glucose

```{r}
a<-func_Accs_no_proteins(corr_gluc_all,df_uniprotIDs)

enr_gluc_corr_path<-func_path_enr(a)
enr_gluc_corr_bp<-func_path_enr(a,"BP")
enr_gluc_corr_cc<-func_path_enr(a,"CC")
enr_gluc_corr_mf<-func_path_enr(a,"MF")

paths_gluc_corr_path.enr <- func_extract_relevant_paths(enr_gluc_corr_path)
paths_gluc_corr_bp <- func_extract_relevant_paths(enr_gluc_corr_bp)
paths_gluc_corr_cc <- func_extract_relevant_paths(enr_gluc_corr_cc)
paths_gluc_corr_mf <- func_extract_relevant_paths(enr_gluc_corr_mf)

# Gene Ontology
taxa_names<-GetNamesTaxa(a)
PlotGenesNetwork(taxa_names)
PlotEnrichedGO(a)

protein_info<-GetProteinGOInfo(a)
PlotGOAll(protein_info)
PlotGoInfo(protein_info)

PlotEnrichedPathways(a)
```

#### Insulin

```{r}
a<-func_Accs_no_proteins(corr_ins_all,df_uniprotIDs)

enr_ins_corr_path<-func_path_enr(a)
enr_ins_corr_bp<-func_path_enr(a,"BP")
enr_ins_corr_cc<-func_path_enr(a,"CC")
enr_ins_corr_mf<-func_path_enr(a,"MF")

paths_ins_corr_path.enr <- func_extract_relevant_paths(enr_ins_corr_path)
paths_ins_corr_bp <- func_extract_relevant_paths(enr_ins_corr_bp)
paths_ins_corr_cc <- func_extract_relevant_paths(enr_ins_corr_cc)
paths_ins_corr_mf <- func_extract_relevant_paths(enr_ins_corr_mf)

# Gene Ontology
taxa_names<-GetNamesTaxa(a)
PlotGenesNetwork(taxa_names)
PlotEnrichedGO(a)

protein_info<-GetProteinGOInfo(a)
PlotGOAll(protein_info)
PlotGoInfo(protein_info)

PlotEnrichedPathways(a)
```

#### Non_HDL_Chol

```{r}
a<-func_Accs_no_proteins(corr_nonhdlchol_all,df_uniprotIDs)

enr_nonhdlchol_corr_path<-func_path_enr(a)
enr_nonhdlchol_corr_bp<-func_path_enr(a,"BP")
enr_nonhdlchol_corr_cc<-func_path_enr(a,"CC")
enr_nonhdlchol_corr_mf<-func_path_enr(a,"MF")

paths_nonhdlchol_corr_path.enr <- func_extract_relevant_paths(enr_nonhdlchol_corr_path)
paths_nonhdlchol_corr_bp <- func_extract_relevant_paths(enr_nonhdlchol_corr_bp)
paths_nonhdlchol_corr_cc <- func_extract_relevant_paths(enr_nonhdlchol_corr_cc)
paths_nonhdlchol_corr_mf <- func_extract_relevant_paths(enr_nonhdlchol_corr_mf)

# Gene Ontology
taxa_names<-GetNamesTaxa(a)
PlotGenesNetwork(taxa_names)
PlotEnrichedGO(a)

protein_info<-GetProteinGOInfo(a)
PlotGOAll(protein_info)
PlotGoInfo(protein_info)

PlotEnrichedPathways(a)
```

#### Triglycerides

```{r}
a<-func_Accs_no_proteins(corr_tg_all,df_uniprotIDs)

enr_tg_corr_path<-func_path_enr(a)
enr_tg_corr_bp<-func_path_enr(a,"BP")
enr_tg_corr_cc<-func_path_enr(a,"CC")
enr_tg_corr_mf<-func_path_enr(a,"MF")

paths_tg_corr_path.enr <- func_extract_relevant_paths(enr_tg_corr_path)
paths_tg_corr_bp <- func_extract_relevant_paths(enr_tg_corr_bp)
paths_tg_corr_cc <- func_extract_relevant_paths(enr_tg_corr_cc)
paths_tg_corr_mf <- func_extract_relevant_paths(enr_tg_corr_mf)

# Gene Ontology
taxa_names<-GetNamesTaxa(a)
PlotGenesNetwork(taxa_names)
PlotEnrichedGO(a)

protein_info<-GetProteinGOInfo(a)
PlotGOAll(protein_info)
PlotGoInfo(protein_info)

PlotEnrichedPathways(a)
```


# Save Paths in RDA

```{r}
# Save all paths obtained
pathways<-mget(ls(pattern = "^paths"))

setwd("./SCRIPTS/Proteomics Analysis")
save(pathways, file="RDA_pathways_proteomics.rda")
```
