==> SCRIPTS/Proteomics Analysis/05_Proteomics_Significant_Pathways_SJ.Rmd <==
---
title: "PF_Proteomics Pathway Enrichment Analysis"
author: 'Uniprot, Annotated Proteins'
date: "SJ"
output:
  html_document:
    toc: yes
    number_sections: true
    toc_depth: 5
  pdf_document:
    toc: yes
    number_sections: true
    toc_depth: 5
urlcolor: blue
linkcolor: red
---

```{r global_options, include=FALSE}
rm(list = ls())
# if (!requireNamespace("BiocManager", quietly = TRUE))
#     install.packages("BiocManager")
# BiocManager::install("EnhancedVolcano")
library(EnhancedVolcano)
library(dplyr)
library(forecast)
library(VennDiagram)
library(readxl) # Input file is xlsx (not csv)
library(UniprotR)
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, comment=NA, warning=FALSE,
                      fig.width = 6, fig.asp = 0.8, out.width = "80%")
#knitr::opts_knit$set(root.dir = "G:/My Drive/Prolonged_Fasting_Study")
knitr::opts_knit$set(root.dir = "../..")
theme_update(plot.title = element_text(hjust = 0.5)) # center plots
```


# Useful functions for getting list results
```{r, message=FALSE}
source("./Functions/Metabolomics functions_SJ_GF.R")
# source("./Functions/Correlation_functions.R")

# To get count of common elements (ordered highest to lowest)
# Although works well for pathways, can be used for other cases.
func_count_common <- function(pathways){
  count_table<-as.data.frame(table(pathways))
  ordered_table<-count_table[order(count_table$Freq, decreasing=T),]
  return (ordered_table)
}
```

# Datasets
```{r}
setwd("./DATA/PROLONGED FASTING/GTAC_somascan/PF_Somalogic_18-045")
df_proteomics=data.frame(read_excel("Cleaned_Proteomics data_AM.xlsx",sheet = "Sheet2_SJedit"))

# Sheet that has Protein name with Xs and dots..
# NOTE: before importing, convert all ",", "+", ":", " ", other punctuation marks to ".". This is important so as to not lead to NAs when merging.
# NOTE: X before numeric value in name taken care of through func_prots_X_numeric
df_uniprotIDs=data.frame(read_excel("Cleaned_Proteomics data_AM.xlsx",sheet = "Uniprot_dict"))
```

# Loaded Significant pathways

```{r}
setwd("./SCRIPTS/Proteomics Analysis")
load("RDA_pathways_proteomics.rda")
# load("RDA_Proteins_FC.rda") # Fold changes in metabolites

# load("RDA_BLEFER_Correlation_Proteins_Biomarkers.rda") # Correlation with timepoints - only BL, EF, ER. 

# load("RDA_All_timepts_Correlation_Proteins_Biomarkers.rda") #Same correlations but with all timepts.

# Note: If W1 included in BLEFER RDA then results almost identical to All-timepts rda.
# Note: If you want to work with W1, then load all-timepts instead of BLEFER.
# Note: BLEFER_corr = corr_biomarker; All-timepts_corr = corr_biomarker_all
```

# Unlist pathways file

```{r, include=FALSE}
list2env(pathways, .GlobalEnv)
rm(pathways)
```

# Common pathways

## Biomarker correlations

### Pathways.Enr

```{r}
paths_corr_path.enr<-rbind(paths_bhb_corr_path.enr,
                           paths_cort_corr_path.enr,
                           paths_igf1_corr_path.enr,
                           paths_t3_corr_path.enr,
                           paths_gluc_corr_path.enr,
                           paths_ins_corr_path.enr,
                           paths_nonhdlchol_corr_path.enr,
                           paths_tg_corr_path.enr)

sig_pathways_corr_path.enr<-func_count_common(paths_corr_path.enr$pathway)
head(sig_pathways_corr_path.enr)
```

### BP

```{r}
paths_corr_bp<-rbind(paths_bhb_corr_bp,
                      paths_cort_corr_bp,
                      paths_igf1_corr_bp,
                      paths_t3_corr_bp,
                      paths_gluc_corr_bp,
                      paths_ins_corr_bp,
                      paths_nonhdlchol_corr_bp,
                      paths_tg_corr_bp)

sig_pathways_corr_bp<-func_count_common(paths_corr_bp$pathway)
head(sig_pathways_corr_bp)
```

### CC

```{r}
paths_corr_cc<-rbind(paths_bhb_corr_cc,
                      paths_cort_corr_cc,
                      paths_igf1_corr_cc,
                      paths_t3_corr_cc,
                      paths_gluc_corr_cc,
                      paths_ins_corr_cc,
                      paths_nonhdlchol_corr_cc,
                      paths_tg_corr_cc)

sig_pathways_corr_cc<-func_count_common(paths_corr_cc$pathway)
head(sig_pathways_corr_cc)
```

### MF

```{r}
paths_corr_mf<-rbind(paths_bhb_corr_mf,
                      paths_cort_corr_mf,
                      paths_igf1_corr_mf,
                      paths_t3_corr_mf,
                      paths_gluc_corr_mf,
                      paths_ins_corr_mf,
                      paths_nonhdlchol_corr_mf,
                      paths_tg_corr_mf)

sig_pathways_corr_mf<-func_count_common(paths_corr_mf$pathway)
head(sig_pathways_corr_mf)
```

## FC intersections

### BP

```{r}
paths_fc_bp<-rbind(paths_blef_blw1_bp,
                   paths_blef_efer_bp)

sig_pathways_fc_bp<-func_count_common(paths_fc_bp$pathway)
head(sig_pathways_fc_bp)
```

### CC

```{r}
paths_fc_cc<-rbind(paths_blef_blw1_cc,
                   paths_blef_efer_cc)

sig_pathways_fc_cc<-func_count_common(paths_fc_cc$pathway)
head(sig_pathways_fc_cc)
```

### MF

```{r}
paths_fc_mf<-rbind(paths_blef_blw1_mf,
                   paths_blef_efer_mf)

sig_pathways_fc_mf<-func_count_common(paths_fc_mf$pathway)
head(sig_pathways_fc_mf)
```

## FC and Biomarker Correlations

### BP

```{r}
paths_fc_corr_bp<-rbind(paths_corr_bp,
                        paths_fc_bp)

sig_pathways_fc_corr_bp<-func_count_common(paths_fc_corr_bp$pathway)
head(sig_pathways_fc_corr_bp)
```

### CC

```{r}
paths_fc_corr_cc<-rbind(paths_corr_cc,
                        paths_fc_cc)

sig_pathways_fc_corr_cc<-func_count_common(paths_fc_corr_cc$pathway)
head(sig_pathways_fc_corr_cc)
```

### MF

```{r}
paths_fc_corr_mf<-rbind(paths_corr_mf,
                        paths_fc_mf)

sig_pathways_fc_corr_mf<-func_count_common(paths_fc_corr_mf$pathway)
head(sig_pathways_fc_corr_mf)
```
